BYT-80 finished connection with backend
