+++ frontend/app/package.json
@@ -22,6 +22,7 @@
     "react-date-range": "^1.4.0",
     "react-dom": "^18.1.0",
     "react-facebook-login": "^4.1.1",
+    "react-query": "^3.39.1",
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
