--- frontend/app/src/Components/Charts/TimeSeriesChart/styles.css
@@ -1,42 +1,0 @@
-.recharts-cartesian-axis-tick {
-  font-size: 0.65rem;
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-.recharts-legend-item {
-  font-size: 0.7rem;
-}
-.recharts-label {
-  font-size: 0.8rem;
-}
-.tooltiptext {
-  visibility: hidden;
-  width: 120px;
-  background-color: grey;
-  opacity: 0.8;
-  color: #fff;
-  text-align: center;
-  border-radius: 6px;
-  padding: 5px 0;
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-  font-size: 0.6rem;
-  z-index: 1;
-  bottom: 150%;
-  left: 50%;
-  margin-left: -60px;
-}
-
-.tooltiptext::after {
-  content: '';
-
-  top: 100%;
-  left: 50%;
-  margin-left: -5px;
-  border-width: 5px;
-  border-style: solid;
-  border-color: black transparent transparent transparent;
-}
-.tooltiptext {
-  visibility: visible;
-}
